Added `inplace` argument
Also updated to use the hosting for the CLI
Moved serilog config to appsettings.config

+++ Episode.Renamer/Program.cs
@@ -1,7 +1,10 @@
 ﻿namespace Episode.Renamer
 {
     using System.CommandLine;
+    using System.CommandLine.Builder;
+    using System.CommandLine.Hosting;
     using System.CommandLine.Invocation;
+    using System.CommandLine.Parsing;
     using System.Linq;
     using System.Threading.Tasks;
     using Microsoft.Extensions.DependencyInjection;
@@ -35,26 +38,42 @@
 
         private static Task<int> Main(string[] args)
         {
-            var host = Microsoft.Extensions.Hosting.Host
-                .CreateDefaultBuilder()
-                .UseSerilog(new LoggerConfiguration()
-                    .WriteTo.Console(formatProvider: System.Globalization.CultureInfo.CurrentCulture)
-                    .MinimumLevel.Information()
-                    .CreateLogger())
-                .Build();
+            return BuildCommandLine()
+                .UseHost(
+                    _ => Microsoft.Extensions.Hosting.Host.CreateDefaultBuilder(),
+                    host => host.UseSerilog((hostBuilderContext, loggerConfiguration) => loggerConfiguration.ReadFrom.Configuration(hostBuilderContext.Configuration)))
+                .UseDefaults()
+                .Build()
+                .InvokeAsync(args.Select(arg => System.Environment.ExpandEnvironmentVariables(arg)).ToArray());
 
-            var root = new RootCommand();
-            root.AddArgument(new Argument<System.IO.DirectoryInfo>("source"));
-            root.AddArgument(new Argument<System.IO.DirectoryInfo>("destination"));
-            root.AddOption(new Option(new[] { "-m", "--move" }, "Moves the files"));
-            root.AddOption(new Option(new[] { "-n", "--dry-run" }, "Don’t actually move/copy any file(s). Instead, just show if they exist and would otherwise be moved/copied by the command."));
+            static CommandLineBuilder BuildCommandLine()
+            {
+                var root = new RootCommand();
+                root.AddArgument(new Argument<System.IO.DirectoryInfo>("source"));
+                root.AddArgument(new Argument<System.IO.DirectoryInfo>("destination"));
+                root.AddOption(new Option<bool>(new[] { "-m", "--move" }, "Moves the files"));
+                root.AddOption(new Option<bool>(new[] { "-r", "--recursive" }, "Recursively searches <SOURCE>"));
+                root.AddOption(new Option<bool>(new[] { "-n", "--dry-run" }, "Don’t actually move/copy any file(s). Instead, just show if they exist and would otherwise be moved/copied by the command."));
+                root.AddOption(new Option<bool>(new[] { "-i", "--inplace" }, "Renames the files in place, rather than to <DESTINATION>."));
 
-            root.Handler = CommandHandler.Create((System.IO.DirectoryInfo source, System.IO.DirectoryInfo destination, bool move, bool dryRun) =>
+                root.Handler = CommandHandler.Create<Microsoft.Extensions.Hosting.IHost, System.IO.DirectoryInfo, System.IO.DirectoryInfo, bool, bool, bool, bool>(Process);
+
+                return new CommandLineBuilder(root);
+            }
+
+            static void Process(
+                Microsoft.Extensions.Hosting.IHost host,
+                System.IO.DirectoryInfo source,
+                System.IO.DirectoryInfo destination,
+                bool move = false,
+                bool recursive = false,
+                bool dryRun = false,
+                bool inplace = false)
             {
                 // search for all files in the source directory
                 var programLogger = host.Services.GetRequiredService<ILogger<Program>>();
                 
-                foreach (var file in source.EnumerateFiles("*.*", new System.IO.EnumerationOptions { RecurseSubdirectories = true, IgnoreInaccessible = true, AttributesToSkip = System.IO.FileAttributes.Hidden }))
+                foreach (var file in source.EnumerateFiles("*.*", new System.IO.EnumerationOptions { RecurseSubdirectories = recursive, IgnoreInaccessible = true, AttributesToSkip = System.IO.FileAttributes.Hidden }))
                 {
                     if (file.Length == 0)
                     {
@@ -66,13 +85,15 @@
 
                     try
                     {
-                        tagLibFile = File.Create(file.FullName);                        
+                        tagLibFile = File.Create(file.FullName);
 
                         if (tagLibFile.GetTag(TagTypes.Apple) is TagLib.Mpeg4.AppleTag appleTag)
                         {
                             if (appleTag.IsMovie())
                             {
-                                var directory = System.IO.Path.Combine(destination.FullName, "Movies").ReplaceAll(GetInvalidPathChars());
+                                var directory = inplace
+                                    ? file.DirectoryName.ReplaceAll(GetInvalidPathChars())
+                                    : System.IO.Path.Combine(destination.FullName, "Movies").ReplaceAll(GetInvalidPathChars());
                                 var fileName = $"{appleTag.Title} ({appleTag.Year}){file.Extension}".ReplaceAll(GetInvalidFileNameChars());
                                 path = new System.IO.FileInfo(System.IO.Path.Combine(directory, fileName));
                             }
@@ -83,7 +104,9 @@
                                 var episodeNumber = appleTag.GetUInt32(EpisodeNumber);
                                 var episodeName = appleTag.Title;
 
-                                var directory = System.IO.Path.Combine(destination.FullName, "TV Shows", showName, $"Season {seasonNumber:00}").ReplaceAll(GetInvalidPathChars());
+                                var directory = inplace 
+                                    ? file.DirectoryName.ReplaceAll(GetInvalidPathChars())
+                                    : System.IO.Path.Combine(destination.FullName, "TV Shows", showName, $"Season {seasonNumber:00}").ReplaceAll(GetInvalidPathChars());
                                 var fileName = $"{showName} - s{seasonNumber:00}e{episodeNumber:00} - {episodeName}{file.Extension}".ReplaceAll(GetInvalidFileNameChars());
 
                                 path = new System.IO.FileInfo(System.IO.Path.Combine(directory, fileName));
@@ -93,14 +116,20 @@
                                 programLogger.LogInformation("Failed to match {File} to either Movie or TV Show", file.Name);
                             }
                         }
+                        else
+                        {
+                            programLogger.LogDebug("Found non 'Apple' format at {File}", file.Name);
+                        }
 
                         tagLibFile.Mode = File.AccessMode.Closed;
                     }
                     catch (UnsupportedFormatException)
                     {
+                        programLogger.LogDebug("Unsupported file - {File}", file.Name);
                     }
                     catch (CorruptFileException)
                     {
+                        programLogger.LogDebug("Corrupt file - {File}", file.Name);
                     }
                     finally
                     {
@@ -129,10 +158,17 @@
                             programLogger.LogInformation("Moving {Source} to {Destination}", file.FullName, path.FullName);
                             if (!dryRun)
                             {
-                                file.CopyTo(path.FullName, true);
-                                if (file.Exists)
+                                if (!path.Exists || inplace)
                                 {
-                                    file.Delete();
+                                    file.MoveTo(path.FullName);
+                                }
+                                else
+                                {
+                                    file.CopyTo(path.FullName, true);
+                                    if (file.Exists)
+                                    {
+                                        file.Delete();
+                                    }
                                 }
                             }
                         }
@@ -146,9 +182,7 @@
                         }
                     }
                 }
-            });
-
-            return root.InvokeAsync(args.Select(arg => System.Environment.ExpandEnvironmentVariables(arg)).ToArray());
+            }
         }
     }
 }
